+++ README.md
@@ -33,7 +33,7 @@
    })
    ```
 
-    ### 와 같은 형식으로 사용한다.
+   ### 와 같은 형식으로 사용한다.
 
 ## 설정 방법
