+++ src/components/ToastContainer.jsx
@@ -3,6 +3,7 @@ import { useEffect, useState } from "react";
 import { createPortal } from "react-dom";
 import EventBus from "../utils/eventBus";
 import AlertToast from "./AlertToast";
+import ConfirmToast from "./ConfirmToast";
 
 const ToastContainer = ({ position, time }) => {
   const [alertToasts, setAlertToasts] = useState([]);
@@ -28,16 +29,6 @@ const ToastContainer = ({ position, time }) => {
     return () => EventBus.unsubscribe();
   }, []);
 
-  const handleConfirm = (id, resolve) => {
-    setConfirmToasts((prevToasts) => prevToasts.filter((t) => t.id !== id));
-    resolve(true);
-  };
-
-  const handleCancel = (id, resolve) => {
-    setConfirmToasts((prevToasts) => prevToasts.filter((t) => t.id !== id));
-    resolve(false);
-  };
-
   return createPortal(
     <>
       <div className={`toast-container ${position}`}>
@@ -52,49 +43,11 @@ const ToastContainer = ({ position, time }) => {
         ))}
 
         {confirmToasts.map((toast) => (
-          <div
+          <ConfirmToast
             key={toast.id}
-            className="toast"
-            style={{
-              width: "100%",
-              position: "fixed",
-              top: 0,
-              left: 0,
-              right: 0,
-              bottom: 0,
-              background: "rgba(0,0,0,0.5)",
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "center",
-            }}
-            onClick={(e) => {
-              if (e.target.classList.contains("toast")) {
-                handleCancel(toast.id, toast.resolve);
-              }
-            }}
-          >
-            <div className="innerBox">
-              {toast.message}
-              <div>
-                <button
-                  onClick={(e) => {
-                    e.preventDefault();
-                    handleConfirm(toast.id, toast.resolve);
-                  }}
-                >
-                  OK
-                </button>
-                <button
-                  onClick={(e) => {
-                    e.preventDefault();
-                    handleCancel(toast.id, toast.resolve);
-                  }}
-                >
-                  Cancel
-                </button>
-              </div>
-            </div>
-          </div>
+            toast={toast}
+            setConfirmToasts={setConfirmToasts}
+          />
         ))}
       </div>
     </>,
